backend/control_app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
channels
Django==5.1.3
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
idna==3.10
python-dotenv==1.0.1
psycopg2-binary
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
